Added Video and CSV Upload Functionality
Once the files have been uploaded, they can be retrieved from the /video and /csvData endpoints. If a csv file is not included in the video upload then one will be generated on the fly. Javascript verifies that the files match. Flask only keeps the latest video and csv; it will discard older versions.